+++ core-ng/src/main/java/core/framework/impl/web/api/EnumDefinition.java
@@ -1,0 +1,13 @@
+package core.framework.impl.web.api;
+
+import core.framework.util.Lists;
+
+import java.util.List;
+
+/**
+ * @author neo
+ */
+public class EnumDefinition {
+    public String name;
+    public List<String> constants = Lists.newArrayList();
+}