+++ core-ng/src/main/java/core/framework/impl/web/api/TypeDefinition.java
@@ -1,0 +1,19 @@
+package core.framework.impl.web.api;
+
+import core.framework.util.Lists;
+
+import java.util.List;
+
+/**
+ * @author neo
+ */
+public class TypeDefinition {
+    public String name;
+    public List<FieldDefinition> fields = Lists.newArrayList();
+
+    public static class FieldDefinition {
+        public String name;
+        public String type;
+        public boolean notNull;
+    }
+}